+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/playback/CastPlayer.kt
@@ -4,6 +4,9 @@ import android.content.Context
 import android.net.Uri
 import android.support.v4.media.session.PlaybackStateCompat
 import android.text.TextUtils
+import androidx.annotation.OptIn
+import androidx.media3.common.PlaybackException
+import androidx.media3.common.util.UnstableApi
 import au.com.shiftyjelly.pocketcasts.models.entity.BaseEpisode
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.UserEpisode
@@ -317,6 +320,7 @@ class CastPlayer(val context: Context, override val onPlayerEvent: (Player, Play
         return null
     }
 
+    @OptIn(UnstableApi::class)
     private fun updatePlaybackState() {
         val remoteMediaClient = remoteMediaClient
         if (remoteMediaClient == null) {
@@ -336,8 +340,20 @@ class CastPlayer(val context: Context, override val onPlayerEvent: (Player, Play
 
         // Convert the remote playback states to media playback states.
         when (status) {
-            MediaStatus.PLAYER_STATE_IDLE -> if (idleReason == MediaStatus.IDLE_REASON_FINISHED) {
-                onPlayerEvent(this, PlayerEvent.Completion(episodeUuid))
+            MediaStatus.PLAYER_STATE_IDLE -> when (idleReason) {
+                MediaStatus.IDLE_REASON_FINISHED -> onPlayerEvent(this, PlayerEvent.Completion(episodeUuid))
+
+                MediaStatus.IDLE_REASON_ERROR -> {
+                    Timber.e("Cast player error for episode $episodeUuid")
+                    val error = PlaybackException("Cast playback error", null, PlaybackException.ERROR_CODE_REMOTE_ERROR)
+                    onPlayerEvent(this, PlayerEvent.PlayerError("Cast playback error", error))
+                }
+
+                MediaStatus.IDLE_REASON_CANCELED -> onPlayerEvent(this, PlayerEvent.PlayerPaused)
+
+                MediaStatus.IDLE_REASON_INTERRUPTED -> Timber.i("Cast playback interrupted for episode $episodeUuid")
+
+                else -> Timber.w("Unknown Cast idle reason $idleReason for episode $episodeUuid")
             }
 
             MediaStatus.PLAYER_STATE_BUFFERING, MediaStatus.PLAYER_STATE_LOADING -> {